# docs/swagger.yaml
basePath: /
definitions:
...
  controller.putSettingsRequest:
    properties:
      app_cors_allowed_origins:
        description: App
        type: string
      jwt_signing_key:
        description: Auth
        minLength: 16
        type: string
      rl_login_limit:
        description: Rate limits (per-endpoint, tenant-scoped overrides)
        type: string
      rl_login_window:
        type: string
      rl_magic_limit:
        type: string
      rl_magic_window:
        type: string
      rl_mfa_limit:
        type: string
      rl_mfa_window:
        type: string
      rl_signup_limit:
        type: string
      rl_signup_window:
        type: string
      rl_sso_limit:
        type: string
      rl_sso_window:
        type: string
      rl_token_limit:
        type: string
      rl_token_window:
        type: string
      scope:
        description: Scope is deprecated and ignored. Kept for backward compatibility
          with older SDKs.
        type: string
      signup_enabled:
        description: Signup & branding
        type: string
      sso_provider:
        type: string
      sso_redirect_allowlist:
        type: string
      sso_state_ttl:
        type: string
      tenant_logo_url:
        type: string
      workos_api_key:
        type: string
      workos_client_id:
        type: string
      workos_client_secret:
        type: string
      workos_default_connection_id:
        type: string
      workos_default_organization_id:
        type: string
    type: object
...
  controller.settingsResponse:
    properties:
      app_cors_allowed_origins:
        description: App
        type: string
      rl_login_limit:
        description: Rate limits (per-endpoint, tenant-scoped overrides)
        type: string
      rl_login_window:
        type: string
      rl_magic_limit:
        type: string
      rl_magic_window:
        type: string
      rl_mfa_limit:
        type: string
      rl_mfa_window:
        type: string
      rl_signup_limit:
        type: string
      rl_signup_window:
        type: string
      rl_sso_limit:
        type: string
      rl_sso_window:
        type: string
      rl_token_limit:
        type: string
      rl_token_window:
        type: string
      signup_enabled:
        description: Signup & branding
        type: string
      sso_provider:
        description: SSO
        type: string
      sso_redirect_allowlist:
        type: string
      sso_state_ttl:
        type: string
      tenant_logo_url:
        type: string
      workos_api_key:
        description: masked
        type: string
      workos_client_id:
        type: string
      workos_client_secret:
        description: masked
        type: string
      workos_default_connection_id:
        type: string
      workos_default_organization_id:
        type: string
    type: object
...
info:
  contact: {}
  description: Central Authentication Service for multi-tenant identity management.
  title: Guard CAS API
  version: "1.0"
paths:
...
  /api/v1/tenants/{id}/settings:
    get:
      description: Returns tenant-scoped settings required for SSO setup (WorkOS)
      parameters:
      - description: Tenant ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.settingsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get tenant settings (subset)
      tags:
      - tenants
    put:
      consumes:
      - application/json
      description: Upserts tenant settings for SSO setup (WorkOS). Only whitelisted
        keys are accepted.
      parameters:
      - description: Tenant ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: settings
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.putSettingsRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upsert tenant settings (subset)
      tags:
      - tenants
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"